Refactor mech2dmanager
